Changed the public interface of GridBagLayoutInfo to use one function to set all alignment or expansion flags and made the flags public. Accessing the flags is currently inconvenient because they are contained in a public member enum, and the enum members cannot be brought into the current scope, so you must always use GridBagLayoutInfo::FOOFLAG to access them.
Changed how GBLM treats maximal sizes. It no longer keeps track of the max size for an entire group, only individual objects. If the maximal size for an object is violated, any expansion thats active on that object will be ignored, and the object will be positioned as if its natural size was being used. The natural size isn't used however, the maximal is.